Moved interaction logic to Chart, Point, Series.
