+++ js/Core/Axis/Axis.js
@@ -286,8 +286,6 @@ var Axis = /** @class */ (function () {
         this.minPixelPadding = void 0;
         this.names = void 0;
         this.offset = void 0;
-        this.oldMax = void 0;
-        this.oldMin = void 0;
         this.options = void 0;
         this.overlap = void 0;
         this.paddedTicks = void 0;
