+++ ts/Core/Axis/TimeTicksInfoObject.d.ts
@@ -1,0 +1,22 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface TimeTicksInfoObject extends Highcharts.TimeNormalizedObject {
+    higherRanks: Array<string>;
+    totalRange: number;
+}
+
+export default TimeTicksInfoObject;
