+++ ts/Series/Line/LineSeries.ts
@@ -2650,6 +2650,8 @@ class LineSeries {
      *
      * */
 
+    public _hasTracking?: boolean;
+
     public _i: number = void 0 as any;
 
     public animationTimeout?: number;
@@ -2698,6 +2700,8 @@ class LineSeries {
 
     public eventsToUnbind: Array<Function> = void 0 as any;
 
+    public halo?: SVGElement;
+
     public hasCartesianSeries?: Chart['hasCartesianSeries'];
 
     public hasRendered?: boolean;
@@ -2738,7 +2742,11 @@ class LineSeries {
 
     public processedYData: (Array<(number|null)>|Array<Array<(number|null)>>) = void 0 as any;
 
+    public selected?: boolean;
+
     public sharedClipKey?: string;
+
+    public stateMarkerGraphic?: SVGElement;
 
     public stickyTracking?: boolean;
 
@@ -2747,6 +2755,10 @@ class LineSeries {
     public tooltipOptions: Highcharts.TooltipOptions = void 0 as any;
 
     public touched?: boolean;
+
+    public tracker?: SVGElement;
+
+    public trackerGroups?: Array<string>;
 
     public userOptions: DeepPartial<SeriesTypeOptions> = void 0 as any;
 
@@ -7122,6 +7134,396 @@ class LineSeries {
                 plotOptions && plotOptions.series && (plotOptions as any).series[optionName],
                 option
             );
+    }
+
+    /**
+     * Runs on mouse over the series graphical items.
+     *
+     * @function Highcharts.Series#onMouseOver
+     * @fires Highcharts.Series#event:mouseOver
+     */
+    public onMouseOver(): void {
+        var series = this,
+            chart = series.chart,
+            hoverSeries = chart.hoverSeries,
+            pointer = chart.pointer;
+
+        pointer.setHoverChartIndex();
+
+        // set normal state to previous series
+        if (hoverSeries && hoverSeries !== series) {
+            hoverSeries.onMouseOut();
+        }
+
+        // trigger the event, but to save processing time,
+        // only if defined
+        if ((series.options.events as any).mouseOver) {
+            fireEvent(series, 'mouseOver');
+        }
+
+        // hover this
+        series.setState('hover');
+
+        /**
+         * Contains the original hovered series.
+         *
+         * @name Highcharts.Chart#hoverSeries
+         * @type {Highcharts.Series|null}
+         */
+        chart.hoverSeries = series;
+    }
+
+    /**
+     * Runs on mouse out of the series graphical items.
+     *
+     * @function Highcharts.Series#onMouseOut
+     *
+     * @fires Highcharts.Series#event:mouseOut
+     */
+    public onMouseOut(): void {
+        // trigger the event only if listeners exist
+        var series = this,
+            options = series.options,
+            chart = series.chart,
+            tooltip = chart.tooltip,
+            hoverPoint = chart.hoverPoint;
+
+        // #182, set to null before the mouseOut event fires
+        chart.hoverSeries = null as any;
+
+        // trigger mouse out on the point, which must be in this series
+        if (hoverPoint) {
+            hoverPoint.onMouseOut();
+        }
+
+        // fire the mouse out event
+        if (series && (options.events as any).mouseOut) {
+            fireEvent(series, 'mouseOut');
+        }
+
+
+        // hide the tooltip
+        if (
+            tooltip &&
+            !series.stickyTracking &&
+            (!tooltip.shared || series.noSharedTooltip)
+        ) {
+            tooltip.hide();
+        }
+
+        // Reset all inactive states
+        chart.series.forEach(function (s): void {
+            s.setState('', true);
+        });
+
+    }
+
+    /**
+     * Set the state of the series. Called internally on mouse interaction
+     * operations, but it can also be called directly to visually
+     * highlight a series.
+     *
+     * @function Highcharts.Series#setState
+     *
+     * @param {Highcharts.SeriesStateValue|""} [state]
+     *        The new state, can be either `'hover'`, `'inactive'`, `'select'`,
+     *        or `''` (an empty string), `'normal'` or `undefined` to set to
+     *        normal state.
+     * @param {boolean} [inherit]
+     *        Determines if state should be inherited by points too.
+     */
+    public setState(
+        state?: (StatesOptionsKey|''),
+        inherit?: boolean
+    ): void {
+        var series = this,
+            options = series.options,
+            graph = series.graph,
+            inactiveOtherPoints = options.inactiveOtherPoints,
+            stateOptions = options.states,
+            lineWidth = options.lineWidth,
+            opacity = options.opacity,
+            // By default a quick animation to hover/inactive,
+            // slower to un-hover
+            stateAnimation = pick(
+                (
+                    (stateOptions as any)[state || 'normal'] &&
+                    (stateOptions as any)[state || 'normal'].animation
+                ),
+                (series.chart.options.chart as any).animation
+            ),
+            attribs,
+            i = 0;
+
+        state = state || '';
+
+        if (series.state !== state) {
+
+            // Toggle class names
+            [
+                series.group,
+                series.markerGroup,
+                series.dataLabelsGroup
+            ].forEach(function (
+                group: (SVGElement|undefined)
+            ): void {
+                if (group) {
+                    // Old state
+                    if (series.state) {
+                        group.removeClass('highcharts-series-' + series.state);
+                    }
+                    // New state
+                    if (state) {
+                        group.addClass('highcharts-series-' + state);
+                    }
+                }
+            });
+
+            series.state = state;
+
+            if (!series.chart.styledMode) {
+
+                if ((stateOptions as any)[state] &&
+                    (stateOptions as any)[state].enabled === false
+                ) {
+                    return;
+                }
+
+                if (state) {
+                    lineWidth = (
+                        (stateOptions as any)[state].lineWidth ||
+                        lineWidth + (
+                            (stateOptions as any)[state].lineWidthPlus || 0
+                        )
+                    ); // #4035
+
+                    opacity = pick(
+                        (stateOptions as any)[state].opacity,
+                        opacity
+                    );
+                }
+
+                if (graph && !graph.dashstyle) {
+                    attribs = {
+                        'stroke-width': lineWidth
+                    };
+
+                    // Animate the graph stroke-width.
+                    graph.animate(
+                        attribs,
+                        stateAnimation
+                    );
+                    while ((series as any)['zone-graph-' + i]) {
+                        (series as any)['zone-graph-' + i].attr(attribs);
+                        i = i + 1;
+                    }
+                }
+
+                // For some types (pie, networkgraph, sankey) opacity is
+                // resolved on a point level
+                if (!inactiveOtherPoints) {
+                    [
+                        series.group,
+                        series.markerGroup,
+                        series.dataLabelsGroup,
+                        series.labelBySeries
+                    ].forEach(function (
+                        group: (SVGElement|undefined)
+                    ): void {
+                        if (group) {
+                            group.animate(
+                                {
+                                    opacity: opacity
+                                },
+                                stateAnimation
+                            );
+                        }
+                    });
+                }
+            }
+        }
+
+        // Don't loop over points on a series that doesn't apply inactive state
+        // to siblings markers (e.g. line, column)
+        if (inherit && inactiveOtherPoints && series.points) {
+            series.setAllPointsToState(state || void 0);
+        }
+    }
+
+    /**
+     * Set the state for all points in the series.
+     *
+     * @function Highcharts.Series#setAllPointsToState
+     *
+     * @private
+     *
+     * @param {string} [state]
+     *        Can be either `hover` or undefined to set to normal state.
+     */
+    public setAllPointsToState(state?: StatesOptionsKey): void {
+        this.points.forEach(function (point): void {
+            if (point.setState) {
+                point.setState(state);
+            }
+        });
+    }
+
+    /**
+     * Show or hide the series.
+     *
+     * @function Highcharts.Series#setVisible
+     *
+     * @param {boolean} [visible]
+     * True to show the series, false to hide. If undefined, the visibility is
+     * toggled.
+     *
+     * @param {boolean} [redraw=true]
+     * Whether to redraw the chart after the series is altered. If doing more
+     * operations on the chart, it is a good idea to set redraw to false and
+     * call {@link Chart#redraw|chart.redraw()} after.
+     *
+     * @fires Highcharts.Series#event:hide
+     * @fires Highcharts.Series#event:show
+     */
+    public setVisible(
+        vis?: boolean,
+        redraw?: boolean
+    ): void {
+        var series = this,
+            chart = series.chart,
+            legendItem = series.legendItem,
+            showOrHide: ('hide'|'show'),
+            ignoreHiddenSeries =
+                (chart.options.chart as any).ignoreHiddenSeries,
+            oldVisibility = series.visible;
+
+        // if called without an argument, toggle visibility
+        series.visible =
+            vis =
+            series.options.visible =
+            series.userOptions.visible =
+            typeof vis === 'undefined' ? !oldVisibility : vis; // #5618
+        showOrHide = vis ? 'show' : 'hide';
+
+        // show or hide elements
+        [
+            'group',
+            'dataLabelsGroup',
+            'markerGroup',
+            'tracker',
+            'tt'
+        ].forEach(function (key: string): void {
+            if ((series as any)[key]) {
+                (series as any)[key][showOrHide]();
+            }
+        });
+
+
+        // hide tooltip (#1361)
+        if (
+            chart.hoverSeries === series ||
+            (chart.hoverPoint && chart.hoverPoint.series) === series
+        ) {
+            series.onMouseOut();
+        }
+
+
+        if (legendItem) {
+            chart.legend.colorizeItem(series, vis);
+        }
+
+
+        // rescale or adapt to resized chart
+        series.isDirty = true;
+        // in a stack, all other series are affected
+        if (series.options.stacking) {
+            chart.series.forEach(function (otherSeries): void {
+                if (otherSeries.options.stacking && otherSeries.visible) {
+                    otherSeries.isDirty = true;
+                }
+            });
+        }
+
+        // show or hide linked series
+        series.linkedSeries.forEach(function (otherSeries): void {
+            otherSeries.setVisible(vis, false);
+        });
+
+        if (ignoreHiddenSeries) {
+            chart.isDirtyBox = true;
+        }
+
+        fireEvent(series, showOrHide);
+
+        if (redraw !== false) {
+            chart.redraw();
+        }
+    }
+
+    /**
+     * Show the series if hidden.
+     *
+     * @sample highcharts/members/series-hide/
+     *         Toggle visibility from a button
+     *
+     * @function Highcharts.Series#show
+     * @fires Highcharts.Series#event:show
+     */
+    public show(): void {
+        this.setVisible(true);
+    }
+
+    /**
+     * Hide the series if visible. If the
+     * [chart.ignoreHiddenSeries](https://api.highcharts.com/highcharts/chart.ignoreHiddenSeries)
+     * option is true, the chart is redrawn without this series.
+     *
+     * @sample highcharts/members/series-hide/
+     *         Toggle visibility from a button
+     *
+     * @function Highcharts.Series#hide
+     * @fires Highcharts.Series#event:hide
+     */
+    public hide(): void {
+        this.setVisible(false);
+    }
+
+
+    /**
+     * Select or unselect the series. This means its
+     * {@link Highcharts.Series.selected|selected}
+     * property is set, the checkbox in the legend is toggled and when selected,
+     * the series is returned by the {@link Highcharts.Chart#getSelectedSeries}
+     * function.
+     *
+     * @sample highcharts/members/series-select/
+     *         Select a series from a button
+     *
+     * @function Highcharts.Series#select
+     *
+     * @param {boolean} [selected]
+     * True to select the series, false to unselect. If undefined, the selection
+     * state is toggled.
+     *
+     * @fires Highcharts.Series#event:select
+     * @fires Highcharts.Series#event:unselect
+     */
+    public select(selected?: boolean): void {
+        var series = this;
+
+        series.selected =
+        selected =
+        this.options.selected = (
+            typeof selected === 'undefined' ?
+                !series.selected :
+                selected
+        );
+
+        if (series.checkbox) {
+            series.checkbox.checked = selected;
+        }
+
+        fireEvent(series, selected ? 'select' : 'unselect');
     }
 
     /** eslint-enable valid-jsdoc */
